[FEAT] SearchHistory post api추가 및 그에 따른 기존 코드 리팩터링

+++ src/main/java/com/example/parking/dto/SearchHistoryDto.java
@@ -9,5 +9,5 @@ import lombok.*;
 @Builder
 @ToString
 public class SearchHistoryDto {
-    private Integer historyId;
+    private Long historyId;
 }
